# More perf improvements, cleanup