--- a/fieldpath.go
+++ b/fieldpath.go
@@ -118,13 +118,12 @@ func (fp *fieldpath) addField() {
 	var name string
 	var i int
 
-	if debugLevel >= 6 {
+	if v(6) {
 		var path string
 		for i := 0; i < len(fp.index)-1; i++ {
 			path += strconv.Itoa(int(fp.index[i])) + "/"
 		}
-
-		_debugfl(6, "Adding field with path: %s%d", path, fp.index[len(fp.index)-1])
+		_debugf("adding field with path: %s%d", path, fp.index[len(fp.index)-1])
 	}
 
 	for i = 0; i < len(fp.index)-1; i++ {
@@ -158,146 +157,104 @@ func newFieldpathHuffman() HuffmanTree {
 // All different fieldops below
 
 func PlusOne(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += 1
 }
 
 func PlusTwo(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += 2
 }
 
 func PlusThree(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += 3
 }
 
 func PlusFour(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += 4
 }
 
 func PlusN(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += int32(r.readUBitVarFP()) + 5
 }
 
 func PushOneLeftDeltaZeroRightZero(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = append(fp.index, 0)
 }
 
 func PushOneLeftDeltaZeroRightNonZero(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 }
 
 func PushOneLeftDeltaOneRightZero(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += 1
 	fp.index = append(fp.index, 0)
 }
 
 func PushOneLeftDeltaOneRightNonZero(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += 1
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 }
 
 func PushOneLeftDeltaNRightZero(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += int32(r.readUBitVarFP())
 	fp.index = append(fp.index, 0)
 }
 
 func PushOneLeftDeltaNRightNonZero(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += int32(r.readUBitVarFP()) + 2
 	fp.index = append(fp.index, int32(r.readUBitVarFP())+1)
 }
 
 func PushOneLeftDeltaNRightNonZeroPack6Bits(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += int32(r.readBits(3)) + 2
 	fp.index = append(fp.index, int32(r.readBits(3))+1)
 }
 
 func PushOneLeftDeltaNRightNonZeroPack8Bits(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += int32(r.readBits(4)) + 2
 	fp.index = append(fp.index, int32(r.readBits(4))+1)
 }
 
 func PushTwoLeftDeltaZero(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 }
 
 func PushTwoLeftDeltaOne(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1]++
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 }
 
 func PushTwoLeftDeltaN(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += int32(r.readUBitVar()) + 2
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 }
 
 func PushTwoPack5LeftDeltaZero(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = append(fp.index, int32(r.readBits(5)))
 	fp.index = append(fp.index, int32(r.readBits(5)))
 }
 
 func PushTwoPack5LeftDeltaOne(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1]++
 	fp.index = append(fp.index, int32(r.readBits(5)))
 	fp.index = append(fp.index, int32(r.readBits(5)))
 }
 
 func PushTwoPack5LeftDeltaN(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += int32(r.readUBitVar()) + 2
 	fp.index = append(fp.index, int32(r.readBits(5)))
 	fp.index = append(fp.index, int32(r.readBits(5)))
 }
 
 func PushThreeLeftDeltaZero(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 }
 
 func PushThreeLeftDeltaOne(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1]++
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
@@ -305,8 +262,6 @@ func PushThreeLeftDeltaOne(r *Reader, fp *fieldpath) {
 }
 
 func PushThreeLeftDeltaN(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += int32(r.readUBitVar()) + 2
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
 	fp.index = append(fp.index, int32(r.readUBitVarFP()))
@@ -314,16 +269,12 @@ func PushThreeLeftDeltaN(r *Reader, fp *fieldpath) {
 }
 
 func PushThreePack5LeftDeltaZero(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = append(fp.index, int32(r.readBits(5)))
 	fp.index = append(fp.index, int32(r.readBits(5)))
 	fp.index = append(fp.index, int32(r.readBits(5)))
 }
 
 func PushThreePack5LeftDeltaOne(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1]++
 	fp.index = append(fp.index, int32(r.readBits(5)))
 	fp.index = append(fp.index, int32(r.readBits(5)))
@@ -331,8 +282,6 @@ func PushThreePack5LeftDeltaOne(r *Reader, fp *fieldpath) {
 }
 
 func PushThreePack5LeftDeltaN(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-1] += int32(r.readUBitVar()) + 2
 	fp.index = append(fp.index, int32(r.readBits(5)))
 	fp.index = append(fp.index, int32(r.readBits(5)))
@@ -340,8 +289,6 @@ func PushThreePack5LeftDeltaN(r *Reader, fp *fieldpath) {
 }
 
 func PushN(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	n := int(r.readUBitVar())
 	fp.index[len(fp.index)-1] += int32(r.readUBitVar())
 
@@ -351,8 +298,6 @@ func PushN(r *Reader, fp *fieldpath) {
 }
 
 func PushNAndNonTopological(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	for i := 0; i < len(fp.index); i++ {
 		if r.readBoolean() {
 			fp.index[i] += r.readVarInt32() + 1
@@ -366,29 +311,21 @@ func PushNAndNonTopological(r *Reader, fp *fieldpath) {
 }
 
 func PopOnePlusOne(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = fp.index[:len(fp.index)-1]
 	fp.index[len(fp.index)-1] += 1
 }
 
 func PopOnePlusN(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = fp.index[:len(fp.index)-1]
 	fp.index[len(fp.index)-1] += int32(r.readUBitVarFP()) + 1
 }
 
 func PopAllButOnePlusOne(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = fp.index[:1]
 	fp.index[len(fp.index)-1] += 1
 }
 
 func PopAllButOnePlusN(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = fp.index[:1]
 	fp.index[len(fp.index)-1] += int32(r.readUBitVarFP()) + 1
 }
@@ -398,36 +335,26 @@ func PopAllButOnePlusNPackN(r *Reader, fp *fieldpath) {
 }
 
 func PopAllButOnePlusNPack3Bits(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = fp.index[:1]
 	fp.index[len(fp.index)-1] += int32(r.readBits(3)) + 1
 }
 
 func PopAllButOnePlusNPack6Bits(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = fp.index[:1]
 	fp.index[len(fp.index)-1] += int32(r.readBits(6)) + 1
 }
 
 func PopNPlusOne(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = fp.index[:len(fp.index)-(int(r.readUBitVarFP()))]
 	fp.index[len(fp.index)-1] += 1
 }
 
 func PopNPlusN(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = fp.index[:len(fp.index)-(int(r.readUBitVarFP()))]
 	fp.index[len(fp.index)-1] += r.readVarInt32()
 }
 
 func PopNAndNonTopographical(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index = fp.index[:len(fp.index)-(int(r.readUBitVarFP()))]
 
 	for i := 0; i < len(fp.index); i++ {
@@ -438,8 +365,6 @@ func PopNAndNonTopographical(r *Reader, fp *fieldpath) {
 }
 
 func NonTopoComplex(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	for i := 0; i < len(fp.index); i++ {
 		if r.readBoolean() {
 			fp.index[i] += r.readVarInt32()
@@ -448,14 +373,10 @@ func NonTopoComplex(r *Reader, fp *fieldpath) {
 }
 
 func NonTopoPenultimatePlusOne(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.index[len(fp.index)-2] += 1
 }
 
 func NonTopoComplexPack4Bits(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	for i := 0; i < len(fp.index); i++ {
 		if r.readBoolean() {
 			fp.index[i] += int32(r.readBits(4)) - 7
@@ -464,7 +385,5 @@ func NonTopoComplexPack4Bits(r *Reader, fp *fieldpath) {
 }
 
 func FieldPathEncodeFinish(r *Reader, fp *fieldpath) {
-	_debugfl(10, "Name: %s", fp.parent.Name)
-
 	fp.finished = true
 }