Mostly reducing allocations and significantly reducing the cost of debugging code:

Reduce huffman allocs:
```
BenchmarkMatch2159568145-OLD 1 11252662856 ns/op 2506655376 B/op 141053247 allocs/op
BenchmarkMatch2159568145-NEW 1  9462694535 ns/op 1921834416 B/op 101196564 allocs/op
                                16%              23%             28%
```

Plus reducing debugging allocs/perf:
```
BenchmarkMatch2159568145-OLD 1 11252662856 ns/op 2506655376 B/op 141053247 allocs/op
BenchmarkMatch2159568145-NEW 1  6145306623 ns/op 1172274816 B/op 35949053 allocs/op
                                45%              53%             74%
```

Test suite runtimes:
```
master    85.206s
perf      46.382s
````